Cell type order is now tracked by state machine, fixing bugs
