+++ frontend/src/Project/service/labels/cellTypesMachine.js
@@ -179,6 +179,7 @@ const createCellTypesMachine = ({ eventBuses, undoRef }) =>
                 CELLS: { actions: 'setCells' },
                 RESTORE: { actions: ['restore', 'updateColorMap'] },
                 SET_FEATURE: { actions: ['setFeature', 'updateColorMap'] },
+                REORDER: { actions: 'reorder' },
               },
             },
           },
@@ -283,6 +284,26 @@ const createCellTypesMachine = ({ eventBuses, undoRef }) =>
               { to: ctx.historyRef }
             ),
           ];
+        }),
+
+        // Reorder the celltypes in the current feature while keeping the other celltypes the same
+        reorder: assign({
+          cellTypes: (ctx, evt) => {
+            const reorder = (list, startIndex, endIndex) => {
+              const result = Array.from(list);
+              const [removed] = result.splice(startIndex, 1);
+              result.splice(endIndex, 0, removed);
+
+              return result;
+            };
+            let cellTypes = ctx.cellTypes.filter((cellType) => cellType.feature !== ctx.feature);
+            let featureCellTypes = ctx.cellTypes.filter(
+              (cellType) => cellType.feature === ctx.feature
+            );
+            let reorderedCellTypes = reorder(featureCellTypes, evt.source, evt.destination);
+            const finalCellTypes = [...cellTypes, ...reorderedCellTypes];
+            return finalCellTypes;
+          },
         }),
 
         // Add one cell to a specified cell type
